refactor : muevo controllers a un solo controlador y arreglos los tests unitarios
